Adds method to update the status of the Empty View.
The updateEmptyStatus() method checks if there is data to display or not, if there is data it hides the empty view, otherwise it will show it. That method is taken directly from the Android source code (https://github.com/android/platform_frameworks_base/blob/master/core/java/android/widget/AdapterView.java#L730-L754).
Then the method is called when the AdapterDataSetObserver gets invalidated (onInvalidated) or the data on it changes (onChanged) before requesting a Layout pass.